requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
pydantic[email]==2.5.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
tokenizers>=0.15.2
git+https://github.com/openai/CLIP.git
tensorflow>=2.16.1
pillow==10.1.0
numpy>=1.26.0
faiss-cpu>=1.8.0
pandas==2.1.3
